requirements.txt
requests~=2.25.1
beautifulsoup4~=4.9.3
jinja2~=3.0.1
lxml
stem
pysocks
aprslib~=0.6.47
paho_mqtt==1.5.0
paramiko~=2.7.2
routeros_api
pyserial~=3.5
html2text~=2020.1.16
pytest~=6.2.4
simplekml~=1.3.5
PyNEC==1.7.3.4 
matplotlib~=3.4.2
flask~=2.0.1
pynput~=1.7.3
numpy
PyYAML~=5.4.1
python-dateutil~=2.8.1
scipy~=1.7.1
librouteros~=3.1.0
pygame~=2.0.1
colorama~=0.4.4
